refactor: User 레거시 코드 제거
