--- backend/src/main/java/com/pickeat/backend/tobe/user/domain/repository/UserRepository.java
@@ -1,21 +1,0 @@
-package com.pickeat.backend.tobe.user.domain.repository;
-
-import com.pickeat.backend.user.domain.User;
-import java.util.List;
-import java.util.Optional;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository("UserRepositoryV2")
-public interface UserRepository extends JpaRepository<User, Long> {
-
-    boolean existsByProviderIdAndProvider(Long providerId, String provider);
-
-    boolean existsByNickname(String nickname);
-
-    Optional<User> findByNickname(String nickname);
-
-    List<User> findByNicknameStartsWith(String nickname);
-
-    List<User> findAllByIdIn(List<Long> userIds);
-}
