--- backend/src/main/java/com/pickeat/backend/tobe/user/ui/UserController.java
@@ -1,41 +1,0 @@
-package com.pickeat.backend.tobe.user.ui;
-
-import com.pickeat.backend.global.auth.annotation.LoginUserId;
-import com.pickeat.backend.tobe.user.application.UserService;
-import com.pickeat.backend.tobe.user.application.dto.UserResponse;
-import com.pickeat.backend.tobe.user.ui.api.UserApiSpec;
-import java.util.List;
-import lombok.RequiredArgsConstructor;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController("UserControllerV2")
-@RequiredArgsConstructor
-@RequestMapping("/api/v2")
-public class UserController implements UserApiSpec {
-
-    private final UserService userService;
-
-    @GetMapping("/users")
-    public ResponseEntity<UserResponse> getUser(@LoginUserId Long userId) {
-        UserResponse response = userService.getById(userId);
-        return ResponseEntity.ok(response);
-    }
-
-    @GetMapping("/rooms/{roomId}/users")
-    public ResponseEntity<List<UserResponse>> getRoomUsers(@PathVariable("roomId") Long roomId) {
-        List<UserResponse> response = userService.getByRoomId(roomId);
-
-        return ResponseEntity.ok(response);
-    }
-
-    @GetMapping("/users/search")
-    public ResponseEntity<List<UserResponse>> getUsers(@RequestParam String nickname) {
-        List<UserResponse> response = userService.searchByNickname(nickname);
-        return ResponseEntity.ok(response);
-    }
-}
