+++ backend/src/main/java/com/pickeat/backend/user/application/UserService.java
@@ -14,7 +14,7 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-@Service
+@Service("UserServiceV2")
 @RequiredArgsConstructor
 @Transactional(readOnly = true)
 public class UserService {
