+++ backend/src/main/java/com/pickeat/backend/user/application/dto/UserResponse.java
@@ -1,8 +1,10 @@
 package com.pickeat.backend.user.application.dto;
 
 import com.pickeat.backend.user.domain.User;
+import io.swagger.v3.oas.annotations.media.Schema;
 import java.util.List;
 
+@Schema(description = "유저 응답", name = "UserResponseV2")
 public record UserResponse(long id, String nickname, Long providerId, String provider) {
 
     public static UserResponse from(User user) {
