+++ backend/src/main/java/com/pickeat/backend/user/domain/repository/UserRepository.java
@@ -4,10 +4,10 @@ import com.pickeat.backend.user.domain.User;
 import java.util.List;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
 
+@Repository("UserRepositoryV2")
 public interface UserRepository extends JpaRepository<User, Long> {
-
-    Optional<User> findByProviderIdAndProvider(Long providerId, String provider);
 
     boolean existsByProviderIdAndProvider(Long providerId, String provider);
