+++ backend/src/main/java/com/pickeat/backend/user/ui/UserController.java
@@ -13,10 +13,9 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
-@Deprecated
-@RestController
+@RestController("UserControllerV2")
 @RequiredArgsConstructor
-@RequestMapping("/api/v1")
+@RequestMapping("/api/v2")
 public class UserController implements UserApiSpec {
 
     private final UserService userService;
