+++ backend/src/test/java/com/pickeat/backend/tobe/user/application/UserServiceTest.java
@@ -13,7 +13,8 @@ import com.pickeat.backend.global.exception.BusinessException;
 import com.pickeat.backend.login.application.dto.request.SignupRequest;
 import com.pickeat.backend.room.domain.Room;
 import com.pickeat.backend.room.domain.RoomUser;
-import com.pickeat.backend.tobe.user.application.dto.UserResponse;
+import com.pickeat.backend.user.application.UserService;
+import com.pickeat.backend.user.application.dto.UserResponse;
 import com.pickeat.backend.user.domain.User;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
