+++ backend/src/test/java/com/pickeat/backend/tobe/wish/application/WishPictureServiceTest.java
@@ -17,12 +17,12 @@ import com.pickeat.backend.room.domain.RoomUser;
 import com.pickeat.backend.room.domain.repository.RoomRepository;
 import com.pickeat.backend.room.domain.repository.RoomUserRepository;
 import com.pickeat.backend.tobe.fixture.WishFixture;
-import com.pickeat.backend.tobe.user.domain.repository.UserRepository;
 import com.pickeat.backend.tobe.wish.application.dto.response.WishPictureResponse;
 import com.pickeat.backend.tobe.wish.domain.Wish;
 import com.pickeat.backend.tobe.wish.domain.repository.WishRepository;
 import com.pickeat.backend.tobe.wish.infrastructure.LocalImageUploadClient;
 import com.pickeat.backend.user.domain.User;
+import com.pickeat.backend.user.domain.repository.UserRepository;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
